+++ ext/lib/Geomatics/DiscCorr.hpp
@@ -162,7 +162,7 @@ namespace gpstk {
          }
 
          ptsdeleted = ptsgood = 0;
-         for(int i=0; i<lines.size(); i++) {
+         for(size_t i=0; i<lines.size(); i++) {
             line = lines[i];
             if(line.empty()) continue;
             // split line into words
